break monolithic event type into 2 types
