+++ tests/Event/BaseEventTest.php
@@ -1,61 +1,15 @@
 <?php
 
-use Basebuilder\Scheduling\Event;
+use Basebuilder\Scheduling\Event\BaseEvent;
 
-class EventTest extends \PHPUnit_Framework_TestCase
+class BaseEventTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * @test
+     * @return BaseEvent|\PHPUnit_Framework_MockObject_MockObject
      */
-    function it_can_compile_the_command()
+    protected function getMock()
     {
-        $event = new Event('php -i');
-
-        $this->assertSame("sh -c 'php -i 1>> '/dev/null' 2>> '/dev/null''", $event->compileCommand());
-    }
-
-    /**
-     * @test
-     */
-    function it_can_switch_user()
-    {
-        $event = new Event('php -i');
-        $event->asUser('foo');
-
-        $this->assertSame("sudo -u foo -- sh -c 'php -i 1>> '/dev/null' 2>> '/dev/null''", $event->compileCommand());
-    }
-
-    /**
-     * @test
-     */
-    function it_can_switch_directory()
-    {
-        $event = new Event('php -i');
-        $event->in('/foo/bar');
-
-        $this->assertSame("cd /foo/bar; sh -c 'php -i 1>> '/dev/null' 2>> '/dev/null''", $event->compileCommand());
-    }
-
-    /**
-     * @test
-     */
-    function it_can_append_output()
-    {
-        $event = new Event('php -i');
-        $event->appendOutput(true);
-
-        $this->assertSame("sh -c 'php -i 1>> '/dev/null' 2>> '/dev/null''", $event->compileCommand());
-    }
-
-    /**
-     * @test
-     */
-    function it_can_overwrite_output()
-    {
-        $event = new Event('php -i');
-        $event->appendOutput(false);
-
-        $this->assertSame("sh -c 'php -i 1> '/dev/null' 2> '/dev/null''", $event->compileCommand());
+        return $this->getMockBuilder(BaseEvent::class)->getMockForAbstractClass();
     }
 
     /**
@@ -63,7 +17,8 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_can_run_every_minute()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
+
         $event->everyMinute();
 
         $this->assertSame('* * * * * *', (string) $event->getCronExpression());
@@ -74,7 +29,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_can_run_n_minutes()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
         $event->everyNMinutes(5);
 
         $this->assertSame("*/5 * * * * *", (string) $event->getCronExpression());
@@ -85,7 +40,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_can_run_hourly()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
         $event->hourly();
 
         $this->assertSame('0 * * * * *', (string) $event->getCronExpression());
@@ -96,7 +51,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_can_run_on_every_hour()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
         $event->hour(1);
 
         $this->assertSame('* 1 * * * *', (string) $event->getCronExpression());
@@ -107,7 +62,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_can_run_daily()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
         $event->daily();
 
         $this->assertSame('0 0 * * * *', (string) $event->getCronExpression());
@@ -118,7 +73,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_can_run_daily_at_a_specific_time()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
         $event->dailyAt('10:05');
         $this->assertSame('5 10 * * * *', (string) $event->getCronExpression());
 
@@ -131,7 +86,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_can_run_on_specific_days()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
 
         $event->days([1, 2 ,3 ]);
         $this->assertSame('* * * * 1,2,3 *', (string) $event->getCronExpression());
@@ -145,7 +100,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_can_run_on_weekdays_only()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
         $event->weekdays();
 
         $this->assertSame('* * * * 1-5 *', (string) $event->getCronExpression());
@@ -156,7 +111,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_can_run_weekly()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
         $event->weekly();
 
         $this->assertSame('0 0 * * 0 *', (string) $event->getCronExpression());
@@ -167,7 +122,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_can_run_monthly()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
         $event->monthly();
 
         $this->assertSame('0 0 1 * * *', (string) $event->getCronExpression());
@@ -178,7 +133,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_can_run_quarterly()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
         $event->quarterly();
 
         $this->assertSame('0 0 1 */3 * *', (string) $event->getCronExpression());
@@ -189,7 +144,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_can_run_yearly()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
         $event->yearly();
 
         $this->assertSame('0 0 1 1 * *', (string) $event->getCronExpression());
@@ -200,7 +155,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
      */
     function it_allows_for_filtering()
     {
-        $event = new Event('php -i');
+        $event = $this->getMock();
         $event->everyMinute();
 
         $this->assertTrue($event->isDue());
@@ -211,7 +166,7 @@ class EventTest extends \PHPUnit_Framework_TestCase
 
         $this->assertFalse($event->isDue());
 
-        $event = new Event('php -i');
+        $event = $this->getMock();
         $event->everyMinute();
 
         $this->assertTrue($event->isDue());
@@ -221,16 +176,5 @@ class EventTest extends \PHPUnit_Framework_TestCase
         });
 
         $this->assertFalse($event->isDue());
-    }
-
-    /**
-     * @group fix
-     */
-    function it_can_prevent_overlapping()
-    {
-        $event = new Event('php -i');
-        $event->preventOverlapping();
-
-        preg_match('/\(touch \w*; php -i; rm \w*\); 1>> /dev/null 2>> /dev/null/', $event->compileCommand());
     }
 }
